requirements.txt
twitter_sdk
flask
pyyaml